requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-oauth-toolkit==3.0.1
django-otp==1.5.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==3.1.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
mypy-extensions==1.0.0
nanoid==2.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
qrcode==8.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
